Avoid building Trezor protobufs
When the Trezor protobufs are built with an older version of the
protobuf lib they generate a bunch of "undefined: proto.ProtoPackageIsVersion3"
errors. Since loomchain doesn't have any use for the trezor/ledger
support in go-ethereum I've refactored things slightly so all that code
isn't built unless the `usbhubs` build tag is specified. These changes
can be reverted when loomchain is updgraded to a newer protobuf version.